v22.6.60231.0

+++ app/Base Application/codeunit/Archivemanagement.C5063.al
@@ -256,7 +256,7 @@ codeunit 5063 ArchiveManagement
         RestoreDocument: Boolean;
         OldOpportunityNo: Code[20];
         IsHandled: Boolean;
-        DoCheck: Boolean;
+        DoCheck, SkipDeletingLinks : Boolean;
     begin
         OnBeforeRestoreSalesDocument(SalesHeaderArchive, IsHandled);
         if IsHandled then
@@ -319,8 +319,10 @@ codeunit 5063 ArchiveManagement
                 OldOpportunityNo := SalesHeader."Opportunity No.";
                 SalesHeader."Opportunity No." := '';
             end;
-            OnRestoreDocumentOnBeforeDeleteSalesHeader(SalesHeader);
-            SalesHeader.DeleteLinks();
+            SkipDeletingLinks := false;
+            OnRestoreDocumentOnBeforeDeleteSalesHeader(SalesHeader, SkipDeletingLinks);
+            if not SkipDeletingLinks then
+                SalesHeader.DeleteLinks();
             SalesHeader.Delete(true);
             OnRestoreDocumentOnAfterDeleteSalesHeader(SalesHeader);
 
@@ -333,6 +335,7 @@ codeunit 5063 ArchiveManagement
             OnRestoreSalesDocumentOnAfterSalesHeaderInsert(SalesHeader, SalesHeaderArchive);
             SalesHeader.TransferFields(SalesHeaderArchive);
             SalesHeader.Status := SalesHeader.Status::Open;
+            OnRestoreSalesDocumentOnBeforeSalesHeaderValidateFields(SalesHeader, SalesHeaderArchive);
             if SalesHeaderArchive."Sell-to Contact No." <> '' then
                 SalesHeader.Validate("Sell-to Contact No.", SalesHeaderArchive."Sell-to Contact No.")
             else
@@ -678,6 +681,7 @@ codeunit 5063 ArchiveManagement
         SalesCommentLine."Line No." := NextLine;
         SalesCommentLine.Date := WorkDate();
         SalesCommentLine.Comment := StrSubstNo(Text004, Format(SalesHeaderArchive."Version No."));
+        OnRestoreSalesLineCommentsOnBeforeInsertSalesCommentLine(SalesCommentLine);
         SalesCommentLine.Insert();
     end;
 
@@ -892,7 +896,7 @@ codeunit 5063 ArchiveManagement
     end;
 
     [IntegrationEvent(false, false)]
-    local procedure OnRestoreDocumentOnBeforeDeleteSalesHeader(var SalesHeader: Record "Sales Header")
+    local procedure OnRestoreDocumentOnBeforeDeleteSalesHeader(var SalesHeader: Record "Sales Header"; var SkipDeletingLinks: Boolean)
     begin
     end;
 
@@ -958,6 +962,16 @@ codeunit 5063 ArchiveManagement
 
     [IntegrationEvent(false, false)]
     local procedure OnBeforeArchSalesDocumentNoConfirm(var SalesHeader: Record "Sales Header"; var IsHandled: Boolean)
+    begin
+    end;
+
+    [IntegrationEvent(false, false)]
+    local procedure OnRestoreSalesLineCommentsOnBeforeInsertSalesCommentLine(var SalesCommentLine: Record "Sales Comment Line")
+    begin
+    end;
+
+    [IntegrationEvent(false, false)]
+    local procedure OnRestoreSalesDocumentOnBeforeSalesHeaderValidateFields(var SalesHeader: Record "Sales Header"; SalesHeaderArchive: Record "Sales Header Archive");
     begin
     end;
 }
